refactored simulation test
