+++ simulator_prototype/tests/simulation/templates/single_cycle_demo_template_full_prog.xls
